+++ test/type_test.rb
@@ -30,6 +30,28 @@ class TypeTestCase < Minitest::Test
     assert_equal :integer, vector.element_type.kind
   end
 
+  def test_size
+    assert_equal "i64 ptrtoint (ptr getelementptr (float, ptr null, i32 1) to i64)", LLVM::Type.float.size.to_s
+    assert_equal "i64 ptrtoint (ptr getelementptr (double, ptr null, i32 1) to i64)", LLVM::Type.double.size.to_s
+    assert_equal "i64 ptrtoint (ptr getelementptr (i42, ptr null, i32 1) to i64)", LLVM::Type.integer(42).size.to_s
+  end
+
+  def test_align
+    assert_equal "i64 ptrtoint (ptr getelementptr ({ i1, i64 }, ptr null, i64 0, i32 1) to i64)", LLVM::Int64.align.to_s
+  end
+
+  def test_from_i
+    assert_raises(ArgumentError) do
+      LLVM::Type.integer(42).from_i(0, 42)
+    end
+
+    assert_equal 'i8 127', LLVM::Int8.from_i(127, true).to_s
+    assert_equal 'i8 127', LLVM::Int8.from_i(127, false).to_s
+    assert_equal 'i8 -128', LLVM::Int8.from_i(-128, true).to_s
+    assert_equal 'i8 poison', LLVM::Int8.from_i(-128, false).to_s
+    assert_equal 'i8 -128', LLVM::Int8.from_i(-128).to_s
+  end
+
   TO_S_TESTS = [
     [LLVM.Struct(), '{}'],
     [LLVM.Struct("test"), '%test = type opaque'],
@@ -74,6 +96,17 @@ class TypeTestCase < Minitest::Test
     [LLVM::Type.function([], LLVM.Void, varargs: true), 'void (...)'],
 
     [LLVM.Struct(LLVM::Int32, LLVM::Type.array(LLVM::Float)), '{ i32, [0 x float] }'],
+
+    [LLVM::Int32.pointer, 'ptr'],
+    [LLVM::Int32.pointer(42), 'ptr addrspace(42)'],
+
+    [LLVM::Type.opaque_struct("hidden"), '%hidden = type opaque'],
+
+    [LLVM.i(1), "i1"],
+    [LLVM.float, "float"],
+    [LLVM.double, "double"],
+    [LLVM.ptr, "ptr"],
+    [LLVM.void, "void"],
   ].freeze
 
   describe "LLVM::Type#to_s" do